+++ src/sections/solana-summit/speaker-grid.tsx
@@ -65,8 +65,11 @@ const speakerCards: SpeakerCard[] = [
     surname: "Gencel",
     company: "Solflare",
     role: "Co-Founder",
-    image: "/images/summit-germany/partners/vidor.png",
+    image: "/images/summit-germany/vidor-new.jpg",
     xUrl: "https://x.com/vidor_solflare",
+    imagePosition: "50% 24%",
+    imageScaleClass: "scale-[1.04]",
+    imageHoverScaleClass: "group-hover:scale-[1.1]",
   },
   {
     name: "Jessica",
@@ -257,8 +260,11 @@ const speakerCards: SpeakerCard[] = [
     surname: "G",
     company: "dopamyn.fun",
     role: "Co-Founder",
-    image: "/images/summit-germany/goldie.jpg",
+    image: "/images/summit-germany/goldie-new.jpg",
     xUrl: "https://x.com/innitgoldie",
+    imagePosition: "28% 42%",
+    imageScaleClass: "scale-[1.08]",
+    imageHoverScaleClass: "group-hover:scale-[1.14]",
   },
   {
     name: "Newton",
